feat(f-6): decoupled futures exit execution (paper simulated + live market, guard-blocked) (#427)
* docs: design futures decoupled exit execution (F-6)

* docs: implementation plan for futures exit execution (F-6)

* feat(f-6): PseudoOCO entry_price + PnL recording + pluggable close executor


* feat(f-6): order_router exit-monitor poll task + paper wiring (shadow risk-state)


* feat(f-6): LiveExitExecutor (real market flatten, guard-blocked) + live wiring


* test(f-6): e2e paper entry→stop-exit lifecycle

* feat(f-6): harden live exit reconciliation (critical-log on post-fill log failure) + await_fill float


---------

+++ shared/execution/kis_futures_adapter.py
@@ -144,7 +144,7 @@ class KISFuturesAdapter:
     async def await_fill(
         self,
         order_id: str,
-        timeout_seconds: int,  # noqa: ARG002 — honored by executor internally
+        timeout_seconds: float,  # noqa: ARG002 — honored by executor internally
     ) -> Fill | None:
         """Return the stashed fill (or None if missed). ``timeout_seconds``
         is honored by ``_send_kis_futures_order`` itself; this method is
